% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_test.R
\name{ls_test}
\alias{ls_test}
\title{ls_test}
\usage{
ls_test(model, contrast, comp, sig.level = 0.05, power = 0.9)
}
\arguments{
\item{model}{a model}

\item{contrast}{factor used for least square mean estimates and contrasts.}

\item{comp}{vector containing the two factor levels for use in the power calculations}

\item{sig.level}{signficance level used in power calculations}

\item{power}{power used in power calculations}
}
\value{
A list consisting of the following:
\describe{
\item{estimates}{summary(model)}
\item{ls_means}{least square means table (emmGrid object) using proportional
weighting}
\item{contrasts}{pair-wise contrast table (emmGrid object) against control
using no multiplicity correction}
\item{power}{power.t.test result}
}
}
\description{
Helper function that returns parameter estimates, effect sizes, pair-wise
comparisons and power t-test calculations.
}
