% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.modelj}
\alias{summary.modelj}
\title{Method for printing summary table}
\usage{
\method{summary}{modelj}(object, ...)
}
\arguments{
\item{object}{modelj object}
}
\value{

}
\description{
Method for printing summary table
}
