% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.modelj}
\alias{summary.modelj}
\title{Method for printing summary table}
\usage{
\method{summary}{modelj}(x, ...)
}
\arguments{
\item{x}{modelj object}

\item{...}{}
}
\value{

}
\description{
Method for printing summary table
}
